pyproject.toml
[project]
name = "vehicle-design-train"
version = "0.1.0"
description = "SDXL LoRA fine-tuning for vehicle text-to-image (JSONL annotations + local SDXL base)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.0.0",
    "clip",
    "dashscope>=1.25.16",
    "datasets>=3.0.0",
    "diffusers>=0.32.0,<0.34.0",
    "image-reward>=1.5",
    "numpy>=1.26.0",
    "packaging>=24.0",
    "peft>=0.14.0",
    "pillow>=10.0.0",
    "safetensors>=0.4.0",
    "setuptools>=69.0.0,<81",
    "tensorboard>=2.14.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "tqdm>=4.66.0",
    "transformers>=4.44.0,<4.47.0",
    "trl>=0.12.0",
]

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
clip = { git = "https://github.com/openai/CLIP.git" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vehicle_design_train"]

[project.scripts]
vdt-token-stats = "vehicle_design_train.token_stats:main"
vdt-eval-grid = "vehicle_design_train.eval_grid:main"
vdt-eval-metrics = "vehicle_design_train.eval_metrics:main"
vdt-train-sdxl-grpo = "vehicle_design_train.train_sdxl_grpo:main"
